Split JSON writing out into an implicit class
